move FlipperClient to core package
