+++ flipperandroidnoop/src/main/java/com/facebook/flipper/android/AndroidFlipperClient.kt
@@ -1,8 +1,10 @@
 package com.facebook.flipper.android
 
 import android.content.Context
+import com.facebook.flipper.core.FlipperClient
 
 object AndroidFlipperClient {
     @Synchronized
-    fun getInstance(context: Context): FlipperClient = FlipperClient()
+    fun getInstance(context: Context): FlipperClient =
+        FlipperClient()
 }
